rip CDs

+++ src/Disk_Drive/ripDiskWrapper.sh
@@ -13,10 +13,13 @@ ERR_NONE=0
 
 function printUsage
 {
-	echo "$0 <Disk_Drive device id> <Media_Plugin device id> <target file name> <source device> <disk type> <ownerID>";
+	echo "$0 <Disk_Drive device id> <Media_Plugin device id> <target file name> <source device> <disk type> <ownerID> ['<tracklist>']";
+	echo "tracklist format: t1,name1[|t2,name2[|t3,name3]...]"
+	echo "  t1 = track number"
+	echo "  name1 = file name track is to be saved as"
 }
 
-[ "$#" -ne "6" ] && printUsage && exit;
+[ "$#" -lt "6" ] && printUsage && exit;
 
 diskDriveDeviceID=$1;
 mediaPluginDeviceID=$2;
@@ -24,9 +27,11 @@ targetFileName=$3;
 sourceDevice=$4;
 diskType=$5;
 ownerID=$6;
+trackList="$7";
 
 # Disk type possbile values:
 #	0 DISCTYPE_CD_AUDIO 
+#	1 DISCTYPE_CD
 #   2 DISCTYPE_DVD_VIDEO 
 #	6 DISCTYPE_CD_MIXED 
 #	7 DISCTYPE_CD_VCD
@@ -36,7 +41,13 @@ echo "Ripping $sourceDevice to \"$targetFileName\" with a disk of type $diskType
 command="";
 result=$ERR_NONE;
 case $diskType in 
-	2)	command="nice -n 15 dd if=$sourceDevice of=\"/home/public/data/movies/$targetFileName.in-progress-dvd\" bs=10M";;
+	2)
+		command="nice -n 15 dd if=$sourceDevice of=\"/home/public/data/movies/$targetFileName.in-progress-dvd\" bs=10M"
+	;;
+	0|1)
+		Dir="/home/public/data/music/$targetFileName"
+		command="nice -n 15 cdparanoia -d $sourceDevice \$Track - | flac -o $Dir/\$FileName.in-progress-flac -"
+	;;
 	*)	result=$ERR_NOT_SUPPORTED_YET;;
 esac
 
@@ -46,10 +57,25 @@ if [ "$command" == "" ]; then
 	exit;
 fi
 
-if eval $command; then 
-	/usr/pluto/bin/MessageSend dcerouter $diskDriveDeviceID $mediaPluginDeviceID 2 35 20 $ERR_SUCCESS 35 "$targetFileName";
-	mv -f "/home/public/data/movies/$targetFileName.in-progress-dvd" "/home/public/data/movies/$targetFileName.dvd";
-else
-	/usr/pluto/bin/MessageSend dcerouter $diskDriveDeviceID $mediaPluginDeviceID 2 35 20 $ERR_RESULT_FAILURE 35 "$targetFileName";
-	rm "/home/public/data/movies/$targetFileName.in-progress-dvd";
+if [[ "$diskType" == 2 ]]; then
+	if eval $command; then
+		/usr/pluto/bin/MessageSend dcerouter $diskDriveDeviceID $mediaPluginDeviceID 2 35 20 $ERR_SUCCESS 35 "$targetFileName";
+		mv -f "/home/public/data/movies/$targetFileName.in-progress-dvd" "/home/public/data/movies/$targetFileName.dvd";
+	else
+		/usr/pluto/bin/MessageSend dcerouter $diskDriveDeviceID $mediaPluginDeviceID 2 35 20 $ERR_RESULT_FAILURE 35 "$targetFileName";
+		rm "/home/public/data/movies/$targetFileName.in-progress-dvd";
+	fi
+elif [[ "$diskType" == 0 || "$diskType" == 1 ]]; then
+	mkdir -p "$Dir"
+	for File in ${trackList//|/ }; do
+		Track=${File%,*}
+		FileName=${File#*,}
+#		if ! nice -n 15 cdparanoia -d $sourceDevice $Track - | flac -o $Dir/$FileName.in-progress-flac -; then
+		if ! eval $command; then
+			/usr/pluto/bin/MessageSend dcerouter $diskDriveDeviceID $mediaPluginDeviceID 2 35 20 $ERR_RESULT_FAILURE 35 "$FileName"
+			exit
+		fi
+		mv $Dir/$FileName.in-progress-flac $Dir/$FileName.flac
+	done
+	/usr/pluto/bin/MessageSend dcerouter $diskDriveDeviceID $mediaPluginDeviceID 2 35 20 $ERR_SUCCESS 35 "$targetFileName"
 fi
